refactor: split connector forms
